+++ src/popup/index.tsx
@@ -4,7 +4,11 @@ import { R2CLogo } from "@r2c/extension/icons";
 import ExtensionTab from "@r2c/extension/popup/ExtensionTab";
 import FirehoseTab from "@r2c/extension/popup/FirehoseTab";
 import ProfileTab from "@r2c/extension/popup/ProfileTab";
-import { ExtensionState, getExtensionState, toggleExtensionExperiment } from "@r2c/extension/shared/ExtensionState";
+import {
+  ExtensionState,
+  getExtensionState,
+  toggleExtensionExperiment
+} from "@r2c/extension/shared/ExtensionState";
 import { getGitHubUserFromStorage } from "@r2c/extension/utils";
 import * as React from "react";
 import "./index.css";
@@ -52,7 +56,7 @@ class Guide extends React.Component<{}, GuideState> {
               />
             }
           />
-          <Tab id="firehose" title="Firehose" panel={<FirehoseTab />} />
+          <Tab id="firehose" title="Activity Feed" panel={<FirehoseTab />} />
           <Tabs.Expander />
           {this.state.currentUser && (
             <Tab